+++ src/main/java/wooteco/subway/domain/fare/policy/FarePolicy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.fare;
+package wooteco.subway.domain.fare.policy;
 
 public interface FarePolicy {
     double calculate(double baseFare);
