+++ src/main/java/wooteco/subway/domain/fare/policy/age/AdultDiscountPolicy.java
@@ -1,7 +1,5 @@
 package wooteco.subway.domain.fare.policy.age;
 
-import wooteco.subway.domain.fare.AgeDiscountPolicy;
-
 public class AdultDiscountPolicy implements AgeDiscountPolicy {
     public AdultDiscountPolicy() {
     }
