+++ src/main/java/wooteco/subway/domain/fare/policy/age/ChildDiscountPolicy.java
@@ -1,7 +1,5 @@
 package wooteco.subway.domain.fare.policy.age;
 
-import wooteco.subway.domain.fare.AgeDiscountPolicy;
-
 public class ChildDiscountPolicy implements AgeDiscountPolicy {
     private static final int DEDUCTION_FARE = 350;
     private static final double CHILD_DISCOUNT_RATE = 0.5;
