+++ src/main/java/wooteco/subway/domain/fare/policy/age/TeenagerDiscountPolicy.java
@@ -1,7 +1,5 @@
 package wooteco.subway.domain.fare.policy.age;
 
-import wooteco.subway.domain.fare.AgeDiscountPolicy;
-
 public class TeenagerDiscountPolicy implements AgeDiscountPolicy {
     private static final int DEDUCTION_FARE = 350;
     private static final double TEENAGER_DISCOUNT_RATE = 0.8;
