+++ src/main/java/wooteco/subway/domain/fare/policy/distance/DistancePolicy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.fare;
+package wooteco.subway.domain.fare.policy.distance;
 
 public interface DistancePolicy {
     int getFare(int distance);
