+++ src/main/java/wooteco/subway/domain/fare/policy/distance/UnderTenKMPolicy.java
@@ -1,7 +1,5 @@
 package wooteco.subway.domain.fare.policy.distance;
 
-import wooteco.subway.domain.fare.DistancePolicy;
-
 public class UnderTenKMPolicy implements DistancePolicy {
     static final int BASE_FEE = 1250;
