+++ src/test/java/wooteco/subway/domain/fare/AgeDiscountPolicyTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.fare.policy.age.AgeDiscountPolicy;
 
 class AgeDiscountPolicyTest {
